fiz: optomize \n\n
Had to remove puppCall becasue it was taking to create new browsers everytime.
This mean I had to add a check to Crawler.crawl() itself to check out of the loop and close the browser.

It's much faster, but there's still a long wasy to go before any actualy features. Though It's really cool and I'm enjoying making it right now.